Refactor CloudFirst-Storage package
