+++ cloudlink-cloudfirst-storage/src/main/java/com/gluonhq/samples/cloudfirst/views/EditionPresenter.java
@@ -24,17 +24,17 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.cloudfirst.views;
+package com.gluonhq.samples.cloudfirst.views;
 
 import com.gluonhq.charm.glisten.afterburner.GluonPresenter;
 import com.gluonhq.charm.glisten.animation.BounceInRightTransition;
 import com.gluonhq.charm.glisten.control.AppBar;
 import com.gluonhq.charm.glisten.mvc.View;
 import com.gluonhq.charm.glisten.visual.MaterialDesignIcon;
-import com.gluonhq.cloudfirst.CombinedStorage;
-import com.gluonhq.cloudfirst.model.Model;
-import com.gluonhq.cloudfirst.model.Note;
-import com.gluonhq.cloudfirst.service.Service;
+import com.gluonhq.samples.cloudfirst.CloudFirstStorage;
+import com.gluonhq.samples.cloudfirst.model.Model;
+import com.gluonhq.samples.cloudfirst.model.Note;
+import com.gluonhq.samples.cloudfirst.service.Service;
 import javafx.beans.binding.Bindings;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -44,7 +44,7 @@ import javafx.scene.control.TextField;
 import javax.inject.Inject;
 import java.util.ResourceBundle;
 
-public class EditionPresenter extends GluonPresenter<CombinedStorage> {
+public class EditionPresenter extends GluonPresenter<CloudFirstStorage> {
 
     @Inject private Service service;
