+++ cloudlink-cloudfirst-storage/src/main/resources/com/gluonhq/samples/cloudfirst/views/notes.css
